refactor(typescript): rename getTypeParametersString
